admin/requirements.txt
Flask
mysql-connector
requests
